refactor(sdk): rename version_compat to framework_compat
